requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 3D Processing
trimesh>=3.21.0
pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
dash>=2.9.0

# Genetic Algorithms
deap>=1.4.0

# Utilities
opencv-python>=4.7.0
tensorboard>=2.12.0

# OpenSCAD integration
solidpython2>=2.0.0